requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
gunicorn
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
Flask==3.1.2
fonttools==4.61.1
fpdf2==2.8.7
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
img2pdf==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
magic-filter==1.0.12
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.6.4
ocrmypdf==17.3.0
packaging==25.0
pdfminer.six==20260107
pefile==2023.2.7
pi_heif==1.3.0
pikepdf==10.3.0
pillow==12.1.1
pluggy==1.6.0
propcache==0.3.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMuPDF==1.27.1
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-telegram-bot==22.4
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.5
rich==14.3.3
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uharfbuzz==0.53.3
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.1.1
yarl==1.20.1

